+++ src-docs/src/views/theme/color_mode/inverse.tsx
@@ -1,58 +1,15 @@
-import React, { FunctionComponent, ReactNode, useState } from 'react';
+import React from 'react';
 import {
   EuiIcon,
   EuiSpacer,
   EuiText,
   EuiThemeProvider,
-  useEuiTheme,
-  EuiButtonGroup,
   EuiPanel,
-  EuiThemeColorMode,
-  EuiHorizontalRule,
 } from '../../../../../src';
 
-const Box: FunctionComponent<{ children: ReactNode }> = ({ children }) => {
-  const { euiTheme } = useEuiTheme();
-
-  return (
-    <EuiText
-      css={{
-        background: euiTheme.colors.lightShade,
-        padding: euiTheme.size.xl,
-        color: euiTheme.colors.text,
-      }}
-    >
-      <p>{children}</p>
-    </EuiText>
-  );
-};
-
 export default () => {
-  const boxColorButtons = [
-    {
-      id: 'light',
-      label: 'Light',
-    },
-    {
-      id: 'dark',
-      label: 'Dark',
-    },
-    {
-      id: 'inverse',
-      label: 'Inverse',
-    },
-  ];
-
-  const [boxColorModeSelected, setBoxColorMode] = useState<EuiThemeColorMode>(
-    'light'
-  );
-
-  const onChange = (colorMode: EuiThemeColorMode) => {
-    setBoxColorMode(colorMode);
-  };
-
   return (
-    <div>
+    <>
       <EuiThemeProvider colorMode="light">
         <EuiPanel>
           <EuiText color="default">
@@ -85,24 +42,6 @@ export default () => {
           </EuiText>
         </EuiPanel>
       </EuiThemeProvider>
-
-      <EuiHorizontalRule />
-
-      <EuiButtonGroup
-        legend="Change color mode"
-        options={boxColorButtons}
-        idSelected={boxColorModeSelected}
-        onChange={(id) => onChange(id as EuiThemeColorMode)}
-      />
-
-      <EuiSpacer />
-
-      <EuiThemeProvider colorMode={boxColorModeSelected}>
-        <Box>
-          <EuiIcon type="faceHappy" /> The colors of this box is in{' '}
-          <strong>{boxColorModeSelected}</strong> color mode
-        </Box>
-      </EuiThemeProvider>
-    </div>
+    </>
   );
 };
